Workers/tests/submissions/Opera/interfaces/DedicatedWorkerGlobalScope/postMessage/second-argument-null-in-array.html

<!--
try {
  postMessage(undefined, [null]);
} catch(e) {
  postMessage(''+e);
}
/*
-->
<!doctype html>
<title>Using [null] in postMessage's second argument</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
(async_test()).step(function() {
  var worker = new Worker('#');
  worker.onmessage = this.step_func(function(e) {
    assert_not_equals(e.data, undefined);
    this.done();
  });
});
</script>
<!--
*/
//-->

